+++ test/Feature/CleanRegex/Match/group/offsets/MatchGroupOffsetStreamTest.php
@@ -198,42 +198,6 @@ class MatchGroupOffsetStreamTest extends TestCase
     /**
      * @test
      */
-    public function shouldGet_remaining_first(): void
-    {
-        // when
-        $first = pattern('(\d+)')->match('Foo 45 65')->remaining(DetailFunctions::equals('65'))->group(1)->offsets()->first();
-
-        // then
-        $this->assertSame(7, $first);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGet_remaining_first_keys(): void
-    {
-        // when
-        $key = pattern('(\d+)')->match('90 60 75 85')->remaining(DetailFunctions::equals('75'))->group(1)->offsets()->keys()->first();
-
-        // then
-        $this->assertSame(0, $key);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldGet_remaining_all_keys(): void
-    {
-        // when
-        $key = pattern('(\d+)')->match('90 60 75 85')->remaining(Functions::oneOf(['60', '85']))->group(1)->offsets()->keys()->all();
-
-        // then
-        $this->assertSame([0, 1], $key);
-    }
-
-    /**
-     * @test
-     */
     public function shouldMapFirst()
     {
         // when