+++ test/Feature/CleanRegex/Match/group/stream/MatchPatternTest.php
@@ -47,42 +47,6 @@ class MatchPatternTest extends TestCase
     /**
      * @test
      */
-    public function shouldReturn_keys_all()
-    {
-        // when
-        $groups = pattern('\d+(?<unit>kg|[cm]?m)')
-            ->match('15mm 12kg 16m 17cm 27kg')
-            ->remaining(Functions::oneOf(['12kg', '16m', '27kg']))
-            ->group('unit')
-            ->stream()
-            ->keys()
-            ->all();
-
-        // then
-        $this->assertSame([0, 1, 2], $groups);
-    }
-
-    /**
-     * @test
-     */
-    public function shouldReturn_keys_first()
-    {
-        // when
-        $groups = pattern('\d+(?<unit>kg|[cm]?m)')
-            ->match('15mm 12kg 16m 17cm 27kg')
-            ->remaining(DetailFunctions::equals('16m'))
-            ->group('unit')
-            ->stream()
-            ->keys()
-            ->first();
-
-        // then
-        $this->assertSame(0, $groups);
-    }
-
-    /**
-     * @test
-     */
     public function shouldGet_map()
     {
         // when